Recommit

+++ src/user/command_executor.cpp
@@ -28,7 +28,7 @@ void CommandExecutor::Execute_Command(std::vector<Command> commands, const kiv_o
 		kiv_os::THandle handle_in = stdin_handle;
 		kiv_os::THandle handle_out = stdout_handle;
 		if (command.has_input_file) {
-			kiv_os_rtl::Open_File(command.input_filename, static_cast<uint8_t>(kiv_os::NFile_Attributes::System_File), kiv_os::NOpen_File::fmOpen_Always, handle_in);
+			kiv_os_rtl::Open_File(command.input_filename, kiv_os::NFile_Attributes::System_File, kiv_os::NOpen_File::fmOpen_Always, handle_in);
 			if (handle_in == invalid_file_handle) {
 				kiv_os_rtl::Write_File(stdout_handle, ERROR_MSG_CANT_OPEN_FILE.data(), ERROR_MSG_CANT_OPEN_FILE.size(), chars_written);
 				kiv_os_rtl::Exit(2);
@@ -37,7 +37,7 @@ void CommandExecutor::Execute_Command(std::vector<Command> commands, const kiv_o
 		}
 
 		if (command.has_output_file) {
-			kiv_os_rtl::Open_File(command.output_filename, static_cast<uint8_t>(kiv_os::NFile_Attributes::System_File), static_cast<kiv_os::NOpen_File>(0), handle_out);
+			kiv_os_rtl::Open_File(command.output_filename, kiv_os::NFile_Attributes::System_File, static_cast<kiv_os::NOpen_File>(0), handle_out);
 			if (handle_in == invalid_file_handle) {
 				kiv_os_rtl::Write_File(stdout_handle, ERROR_MSG_CANT_OPEN_FILE.data(), ERROR_MSG_CANT_OPEN_FILE.size(), chars_written);
 				kiv_os_rtl::Exit(2);
